requirements.txt
dj-database-url==0.5.0
gunicorn==19.9.0
psycopg2-binary==2.7.7
whitenoise==4.1.2
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.7
django-filter==2.1.0
djangorestframework==3.9.1
drf-writable-nested==0.5.1
idna==2.8
pytz==2018.9
requests==2.21.0
urllib3==1.24.1